useStudyMaterials query separated into two different queries
